+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/arg_parser.go
@@ -23,11 +23,16 @@ const (
 	serviceIdArgName     = "service_id"
 	serviceConfigArgName = "config"
 	defineFactArgName    = "define_fact"
+	requestArgName       = "request"
 
 	containerImageNameKey = "image"
 	factNameArgName       = "fact_name"
 	usedPortsKey          = "ports"
 	// TODO remove this when we have the Portal as this is a temporary hack to meet the NEAR use case
+	serviceIdKey   = "service_id"
+	contentTypeKey = "content_type"
+	bodyKey        = "body"
+
 	publicPortsKey                 = "public_ports"
 	entryPointArgsKey              = "entrypoint"
 	cmdArgsKey                     = "cmd"
@@ -38,6 +43,8 @@ const (
 	requestMethodEndpointKey       = "method"
 	fieldExtractorKey              = "field_extractor"
 	privateIPAddressPlaceholderKey = "private_ip_address_placeholder"
+
+	httpRequestExtractorsKey = "extract"
 
 	portNumberKey   = "number"
 	portProtocolKey = "protocol"
@@ -123,42 +130,47 @@ func ParseHttpRequestFactRecipe(serviceConfig *starlarkstruct.Struct) (*kurtosis
 	}
 }
 
-func ParseHttpRequestRecipe(serviceConfig *starlarkstruct.Struct) (*recipe.HttpRequestRecipe, *startosis_errors.InterpretationError) {
-	serviceId, interpretationErr := extractStringValue(serviceConfig, "service_id", defineFactArgName)
+func ParseHttpRequestRecipe(recipeConfig *starlarkstruct.Struct) (*recipe.HttpRequestRecipe, *startosis_errors.InterpretationError) {
+	serviceId, interpretationErr := extractStringValue(recipeConfig, serviceIdKey, requestArgName)
 	if interpretationErr != nil {
 		return nil, interpretationErr
 	}
 
-	portId, interpretationErr := extractStringValue(serviceConfig, portIdKey, defineFactArgName)
+	portId, interpretationErr := extractStringValue(recipeConfig, portIdKey, requestArgName)
 	if interpretationErr != nil {
 		return nil, interpretationErr
 	}
 
-	endpoint, interpretationErr := extractStringValue(serviceConfig, requestEndpointKey, defineFactArgName)
+	endpoint, interpretationErr := extractStringValue(recipeConfig, requestEndpointKey, requestArgName)
 	if interpretationErr != nil {
 		return nil, interpretationErr
 	}
 
-	method, interpretationErr := extractStringValue(serviceConfig, requestMethodEndpointKey, defineFactArgName)
+	method, interpretationErr := extractStringValue(recipeConfig, requestMethodEndpointKey, requestArgName)
+	if interpretationErr != nil {
+		return nil, interpretationErr
+	}
+
+	extractors, interpretationErr := parseHttpRequestExtractors(recipeConfig)
 	if interpretationErr != nil {
 		return nil, interpretationErr
 	}
 
 	if method == getRequestMethod {
-		builtConfig := recipe.NewGetHttpRequestRecipe(service.ServiceID(serviceId), portId, endpoint)
+		builtConfig := recipe.NewGetHttpRequestRecipe(service.ServiceID(serviceId), portId, endpoint, extractors)
 		return builtConfig, nil
 	} else if method == postRequestMethod {
-		contentType, interpretationErr := extractStringValue(serviceConfig, "content_type", defineFactArgName)
+		contentType, interpretationErr := extractStringValue(recipeConfig, contentTypeKey, defineFactArgName)
 		if interpretationErr != nil {
 			return nil, interpretationErr
 		}
 
-		body, interpretationErr := extractStringValue(serviceConfig, "body", defineFactArgName)
+		body, interpretationErr := extractStringValue(recipeConfig, bodyKey, defineFactArgName)
 		if interpretationErr != nil {
 			return nil, interpretationErr
 		}
 
-		builtConfig := recipe.NewPostHttpRequestRecipe(service.ServiceID(serviceId), portId, contentType, endpoint, body)
+		builtConfig := recipe.NewPostHttpRequestRecipe(service.ServiceID(serviceId), portId, contentType, endpoint, body, extractors)
 		return builtConfig, nil
 	} else {
 		return nil, startosis_errors.NewInterpretationError("Define fact HTTP method not recognized")
@@ -455,6 +467,19 @@ func parseFilesArtifactMountDirpaths(serviceConfig *starlarkstruct.Struct) (map[
 		return nil, interpretationErr
 	}
 	return filesArtifactMountDirpathsArg, nil
+}
+
+func parseHttpRequestExtractors(recipe *starlarkstruct.Struct) (map[string]string, *startosis_errors.InterpretationError) {
+	_, err := recipe.Attr(httpRequestExtractorsKey)
+	//an error here means that no argument was found which is alright as this is an optional
+	if err != nil {
+		return map[string]string{}, nil
+	}
+	httpRequestExtractorsArg, interpretationErr := extractMapStringStringValue(recipe, httpRequestExtractorsKey, requestArgName)
+	if interpretationErr != nil {
+		return nil, interpretationErr
+	}
+	return httpRequestExtractorsArg, nil
 }
 
 func parsePrivateIPAddressPlaceholder(serviceConfig *starlarkstruct.Struct) (string, *startosis_errors.InterpretationError) {